fix(RouteList): adjust page size for screen height
Load more elements on larger screens to ensure the entire height is
filled and the intersection observer is off screen. If after the
elements are loaded the intersection observer is still on screen it will
never be triggered.

This bug only affected large monitors on the smallest zoom levels, but
the changes mean devices with smaller screens will load fewer elements
too, which might improve performance/load times.