ci: comment with preview URL immediately after deploy (#350)
This means you see the preview URL 20-40 seconds earlier depending on
how long it takes to capture the screenshots.

Unfortunately the comment on pull request action doesn't have a mode to
not create/update the comment if one with the same tag already exists.
This means it will update an existing comment on redeploy, removing the
screenshots, before taking new ones and updating the comment.