fix: RouteActivity suspense not blocking any longer (#474)
suspense works in RouteActivity now. `keyed` makes Solid re-render the
component from scratch, so in this case the `RouteActivity` is
re-rendered from scratch. i think the re-use between navigations was
causing issues with resources in `RouteActivity`

we need `keyed` at the `dongleId` above for a similar reason.

i think it's technically an escape hatch, but we can spend time later on
removing this.

need to rework RouteStatistics, engaged shows NaN% on immediate load


https://github.com/user-attachments/assets/bb0670c6-a869-4c22-a5c6-0be73582c4ac

---------

Co-authored-by: Shane Smiskol <shane@smiskol.com>