feat(PairActivity): add 'grant' permission screen
Check the camera permission status before opening the QR camera and
prompt the user for permission if necessary. If permission is denied, we
can show a message telling the user to check their browser settings.

Firefox is different from Chrome and doesn't tell us explicitly if
permission is denied. I've "stress tested" switching permissions back
and forth on Firefox and Chrome and it does seem to open the QR scanner
correctly when permission is granted.

- [x] Tested on Chrome and Firefox
- [ ] Tested on Android
- [ ] Tested on Safari/iOS